pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "diffusion-privacy-attacks"
version = "0.1.0"
description = "Baseline memorization attacks for diffusion model outputs"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.25",
  "Pillow>=10.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
